pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "auto-writer-wp"
version = "0.1.0"
description = "WordPress 자동 블로그 포스팅 시스템"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
requests = "^2.31.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
apscheduler = "^3.10.4"
openai = "^1.3.0"
pillow = "^10.1.0"
aiofiles = "^23.2.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
aiosqlite = "^0.19.0"
httpx = "^0.28.1"
typer = "^0.9.0"
rich = "^13.7.0"
beautifulsoup4 = "^4.12.2"
python-slugify = "^8.0.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
pytest-cov = "^4.1.0"
pytest-httpx = "^0.35.0"

[tool.poetry.scripts]
auto-writer = "src.apps.cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"